Add SLOW_TEST option
Normally, the tests scaffold the app into a child directory then runs
the tests. This means that the scaffolded app uses the node_modules
directory of the parent. That's why all the deps of a Claytons app are
dev-deps of the Claytons project itself.

This makes it faster, but can mask errors if you forgot to add
something to package.mustache.

The SLOW_TEST flag tells the test runner to do everything in a tmp dir
and run npm install first. It takes forever, because npm, but it's more
robust. Useful for CI environments and pre-release tests.