WIP - Hyperschema experiment
This is currently an abandoned plan. I was thinking it would be good to
expose a hyperchema so that the admin UI could bootstrap itself based
off a standard, rather than convention. However, I can't see a way to
get the entire admin UI out of just what's in the spec. Things like
rel's aren't well specified, the format for paramaters embedded in
url's (/entity/:id) seems pretty ropy, etc.

The big disadvantage of exposing a hyperschema is that as people add
new routes, they'd need to also update the hyperschema to keep in in
sync. That adds the requirement to understand the hyperschema standard
which is a non-trivial barrier to entry. The corresponding gain it
provides is pretty minimal.